mediacapture-depth/MediaStream-depth-audio-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>MediaStream Depth and Audio Streams</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/">
<link rel="help" href="https://w3c.github.io/mediacapture-depth/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<p class="instructions">Use a test device with depth camera(embedded or external) and audio device.</p>
<p class="instructions">When prompted, accept to share your audio and depth streams.</p>

<div id="log"></div>

<script>

promise_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}, audio: true})
  .then(stream => {
      let stream1 = new MediaStream();
      assert_not_equals(stream.id, stream1.id, "Two different MediaStreams have different ids");

      let stream2 = new MediaStream(stream);
      assert_not_equals(stream.id, stream2.id, "A MediaStream constructed from another have different ids");
      let audioTrack1 = stream.getAudioTracks()[0];
      let videoTrack = stream.getVideoTracks()[0];
      assert_equals(audioTrack1, stream2.getAudioTracks()[0], "A MediaStream constructed from another share the same audio track");
      assert_equals(videoTrack, stream2.getVideoTracks()[0], "A MediaStream constructed from another share the same video track");

      let stream3 = new MediaStream([audioTrack1]);
      assert_equals(stream3.getTrackById(audioTrack1.id), audioTrack1, "a non-ended track gets added via the MediaStream constructor");
      let audioTrack2 = audioTrack1.clone();
      audioTrack2.addEventListener("ended", t.unreached_func("ended event should not be fired by MediaStreamTrack.stop()."));
      audioTrack2.stop();
      assert_equals(audioTrack2.readyState, "ended", "a stopped track is marked ended synchronously");

      let stream4 = new MediaStream([audioTrack2, videoTrack]);
      assert_equals(stream4.getTrackById(audioTrack2.id), audioTrack2, "an ended track gets added via the MediaStream constructor");
      assert_equals(stream4.getTrackById(videoTrack.id), videoTrack, "a non-ended track gets added via the MediaStream constructor even if the previous track was ended");

      let stream5 = new MediaStream([audioTrack2]);
      assert_equals(stream5.getTrackById(audioTrack2.id), audioTrack2, "an ended track gets added via the MediaStream constructor");
      assert_false(stream5.active, "a MediaStream created using the MediaStream() constructor whose arguments are lists of MediaStreamTrack objects that are all ended, the MediaStream object MUST be created with its active attribute set to false");
  });
}, "Tests that a MediaStream Depth constructor follows the algorithm set in the spec");

promise_test(t => {
  let audio, video;
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(video_stream => {
    video = video_stream;
    return navigator.mediaDevices.getUserMedia({audio: true})
    .then(audio_stream => {
      audio = audio_stream;
      audio.getAudioTracks()[0].stop();
      assert_false(audio.active, "audio stream is inactive after stopping its only audio track");
      assert_true(video.active, "video stream is active");
      audio.addTrack(video.getVideoTracks()[0]);
      audio.removeTrack(audio.getAudioTracks()[0]);
    });
  });
}, "Tests that adding a track to an inactive MediaStream is allowed");

</script>
